fix: move themeArgs to separate module
