+++ packages/zero-runtime/src/index.ts
@@ -1,9 +1,4 @@
-import styled from './styled';
+export { default as styled } from './styled';
 
 export { default as sx } from './sx';
 export { default as keyframes } from './keyframes';
-
-export type { ThemeArgs } from './base';
-
-export { styled };
-export default styled;
