+++ packages/zero-runtime/src/keyframes.d.ts
@@ -1,4 +1,5 @@
-import type { CSSProperties, ThemeArgs } from './base';
+import type { CSSProperties } from './base';
+import type { ThemeArgs } from './theme';
 
 type Primitve = string | null | undefined | boolean | number;
