+++ packages/zero-runtime/src/sx.d.ts
@@ -1,4 +1,5 @@
-import type { CSSObjectNoCallback, ThemeArgs } from './base';
+import type { CSSObjectNoCallback } from './base';
+import type { ThemeArgs } from './theme';
 
 export type SxProp = CSSObjectNoCallback | ((themeArgs: ThemeArgs) => CSSObjectNoCallback);
